+++ NEWS.md
@@ -1,5 +1,7 @@
 # vroom (development version)
 
+* The progress bar now only updates every 10 milliseconds
+
 * Disable colors in the progress bar when running in RStudio, so work around
   https://github.com/rstudio/rstudio/issues/4777